pyproject.toml
[tool.poetry]
name = "demetrius"
version = "0.1.0"
description = "ok xd"
authors = ["Yan Regojo <yan.regojo@ensea.fr>, Dimitri Benoit <dimitri.benoit@ensea.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
python-decouple = "3.8"
langchain ="0.0.311"
chromadb = "^0.4.13"
unstructured = "^0.10.19"
pdf2image = "1.16.3"
pypdf2 = "3.0.1"
django-cors-headers = "4.3.0"
django-debug-toolbar = "4.2.0"
django-pdb = "0.6.2"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
tiktoken = "^0.5.1"
openai = "0.28.1"
pypdf = "3.17.0"
nltk = "3.8.1"
channels-redis = "4.1.0"
uvicorn = "0.24.0.post1"
websocket = "0.2.1"
lark = "1.1.8"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
